게시글 목록 가져올 때, 전체 Page 수 추가 및 예외처리

+++ src/main/java/me/r6_search/dto/post/PageMetaDto.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class PageMetaDto {
-    int currentPage;
-    int totalPage;
+    long currentPage;
+    long totalPage;
 }
